Supervisor rename.
