+++ lib/rtsp/session.ex
@@ -28,9 +28,8 @@ defmodule Membrane.Protocol.RTSP.Session do
   Starts and links session process.
 
   Sets following properties of Session:
-    * transport - struct representing module that will be used when executing
-    request and reference used for resolving process controlling transport
-    process
+    * transport - information for executing request over the network for
+    reference see `Membrane.Protocol.RTSP.Transport`
     * url - a base path for requests
     * options - a keyword list that shall be passed when executing request over transport
   """
