+++ test/rtsp/session/supervisor_test.exs
@@ -1,8 +1,7 @@
-defmodule Membrane.Protocol.RTSP.Session.CoupleSupervisorTest do
+defmodule Membrane.Protocol.RTSP.Session.SupervisorTest do
   use ExUnit.Case
 
   alias Membrane.Protocol.RTSP.{Session, Transport}
-  alias Membrane.Protocol.RTSP.Session.CoupleSupervisor
 
   @parsed_uri %URI{
     authority: "domain.com:554",
@@ -16,7 +15,7 @@ defmodule Membrane.Protocol.RTSP.Session.CoupleSupervisorTest do
     test "when initializing returns correct spec when valid arguments were provided" do
       ref = "magic_ref"
       transport = Transport.new(Fake, ref)
-      assert {:ok, {_, children_spec}} = CoupleSupervisor.init([transport, @parsed_uri, []])
+      assert {:ok, {_, children_spec}} = Session.Supervisor.init([transport, @parsed_uri, []])
       assert [session_spec, transport_spec] = children_spec
 
       assert %{id: Session, start: {Session, :start_link, [^transport, @parsed_uri, []]}} =
@@ -28,7 +27,7 @@ defmodule Membrane.Protocol.RTSP.Session.CoupleSupervisorTest do
 
     test "start_link returns an error if invalid uri is provided" do
       assert {:stop, :invalid_url} ==
-               CoupleSupervisor.start_link(Fake, "rtsp://vod/mp4:movie.mov", [])
+               Session.Supervisor.start_link(Fake, "rtsp://vod/mp4:movie.mov", [])
     end
   end
 end
